pyproject.toml
[tool.poetry]
name = "ascender-microservices"
version = "0.1.0"
description = "ascender-microservices"
authors = ["Degeron <degeron0@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ascender-framework = "^1.3.0b5"
aiokafka = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
